requirements/tests.txt
-r base.txt

factory-boy==3.2.1
Faker==13.13.0
pytest==7.1.2
pytest-cov==3.0.0
pytest-network==0.0.1
pytest_factoryboy==2.5.0
pytest-asyncio==0.18.3
pytest-postgresql==4.1.1
psycopg[binary]==3.1.6
requests==2.28.0
boto3==1.24.13
boto3-stubs==1.24.13
moto==3.1.14
black==22.12.0
